refactor: PR2 - Create core package (models and handoff_lifecycle)
- Create handoff/core/ package with models.py and handoff_lifecycle.py
- Move models.py -> core/models.py
- Move handoff_lifecycle.py -> core/handoff_lifecycle.py
- Update all imports in data/, services/, pages/, backup_schema,
  page_models, rulebook, db.py, and tests
- Delete old top-level models.py and handoff_lifecycle.py

Part of project-structure-restructure plan.
