+++ src/handoff/core/__init__.py
@@ -1,0 +1,18 @@
+"""Domain models and lifecycle rules shared by all interfaces."""
+
+from handoff.core.handoff_lifecycle import (
+    get_handoff_close_date,
+    handoff_is_closed,
+    handoff_is_open,
+)
+from handoff.core.models import CheckIn, CheckInType, Handoff, Project
+
+__all__ = [
+    "CheckIn",
+    "CheckInType",
+    "Handoff",
+    "Project",
+    "get_handoff_close_date",
+    "handoff_is_closed",
+    "handoff_is_open",
+]
