+++ src/handoff/data/io.py
@@ -8,8 +8,8 @@ from sqlmodel import select
 
 from handoff.backup_schema import BackupPayload
 from handoff.bootstrap.logging import log_application_action
+from handoff.core.models import CheckIn, Handoff, Project
 from handoff.db import session_context
-from handoff.models import CheckIn, Handoff, Project
 
 
 def import_payload(data_payload: dict[str, Any]) -> None:
