+++ src/handoff/data/projects.py
@@ -5,9 +5,9 @@ from __future__ import annotations
 from loguru import logger
 from sqlmodel import select
 
+from handoff.core.models import Project
 from handoff.data.activity import log_activity
 from handoff.db import session_context
-from handoff.models import Project
 
 
 def create_project(name: str) -> Project:
